requirements.txt
bleak==0.22.3
certifi==2025.10.5
cffi==2.0.0
charset-normalizer==3.4.4
cryptography==46.0.3
dbus-fast==2.44.5
idna==3.11
meshtastic==2.7.3
packaging==24.2
paho-mqtt==2.1.0
protobuf==6.33.0
pycparser==2.23
Pypubsub==4.0.3
pyserial==3.5
PyYAML==6.0.3
requests==2.32.5
tabulate==0.9.0
urllib3==2.5.0
